+++ src/main/java/br/pucrio/tecmf/PACLearningModuleCalculator/service/NeuralNetworkCalculatorService.java
@@ -19,7 +19,7 @@ public class NeuralNetworkCalculatorService implements IPACLearningCalculator {
 
 
     @Override
-    public Integer calculateMinimalSample(Integer accuracy, Integer reliability) {
+    public Integer calculateMinimalSample(Double accuracy, Double reliability) {
         return (int) (1/accuracy*(estimateVCDim()+Math.log(1/reliability)));
     }
 
@@ -44,7 +44,7 @@ public class NeuralNetworkCalculatorService implements IPACLearningCalculator {
      * @return
      */
     @Override
-    public Integer[][] lowerBoundsSamplesBetweenAccuracyAndReliability(Integer[] accuracy, Integer[] reliability, Integer VCDim, Integer range) {
+    public Integer[][] lowerBoundsSamplesBetweenAccuracyAndReliability(Double[] accuracy, Double[] reliability, Integer VCDim, Integer range) {
         return new Integer[0][];
     }
