+++ src/main/java/br/pucrio/tecmf/PACLearningModuleCalculator/service/RandomForestCalculatorService.java
@@ -17,7 +17,7 @@ public RandomForestCalculatorService(Integer features) {
      * @return
      */
     @Override
-    public Integer calculateMinimalSample(Integer accuracy, Integer reliability) {
+    public Integer calculateMinimalSample(Double accuracy, Double reliability) {
 
         return (int) (1/accuracy*(estimateVCDim()+Math.log(1/reliability)));
     }
@@ -41,7 +41,7 @@ public RandomForestCalculatorService(Integer features) {
      * @return
      */
     @Override
-    public Integer[][] lowerBoundsSamplesBetweenAccuracyAndReliability(Integer[] accuracy, Integer[] reliability, Integer VCDim, Integer range) {
+    public Integer[][] lowerBoundsSamplesBetweenAccuracyAndReliability(Double[] accuracy, Double[] reliability, Integer VCDim, Integer range) {
         return new Integer[0][];
     }
